Make clipboard an optional
Initially `pattern.py` was designed to run without any requirements so
it can be quickly used on a remote machine if needed. Adding the
clipboard feature however requires a PyPi package.

This patch makes the clipboard package optional. If it is available on
the system, the DeBruijin code is copied to the clipboard, otherwise
it is just printed to stdout.